Fix integer division in a floating point context warnings
Found with clang-tidy

> src/gui/editors/abstract-metatile-editor.cpp:774:32: warning: result of integer division used in a floating point context; possible loss of precision [bugprone-integer-division]
>                               (tileId / TILESET_WIDTH) / float(TILESET_HEIGHT));
>                                ^
> src/gui/editors/metatile-tileset-editor.cpp:509:52: warning: result of integer division used in a floating point context; possible loss of precision [bugprone-integer-division]
>         const ImVec2 uvSize = ImVec2(1.0f, 1.0f / (TILE_COLLISION_IMAGE_HEIGHT / METATILE_SIZE_PX));
>                                                    ^

+++ src/gui/editors/abstract-metatile-editor.cpp
@@ -770,8 +770,8 @@ void AbstractMetaTileEditorGui::drawCursorTiles(const grid<uint16_t>& tiles, con
                 const auto tileId = *it++;
 
                 if (tileId < N_METATILES) {
-                    ImVec2 uv((tileId % TILESET_WIDTH) / float(TILESET_WIDTH),
-                              (tileId / TILESET_WIDTH) / float(TILESET_HEIGHT));
+                    ImVec2 uv(int(tileId % TILESET_WIDTH) / float(TILESET_WIDTH),
+                              int(tileId / TILESET_WIDTH) / float(TILESET_HEIGHT));
 
                     const ImU32 tint = bounds.contains(point(x, y)) ? Style::tileCursorInBoundsTint : Style::tileCursorOutOfBoundsTint;
